pyproject.toml
[project]
name = "t3"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "click>=8.2.1",
    "click-default-group>=1.2.4",
    "things-py>=0.0.15",
]

[dependency-groups]
dev = [
    "nuitka>=2.7.13",
]

[project.scripts]
t3 = "t3.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/t3"]
